Evervolv fix for CM crash-test commit policy
Corrected error in screen backlight level
This allows for backlight to decrease when changing to dimmer location
when phone is not connected to a dock.  Previouos setting had reversed.

Change-Id: I3f9e44dec8e3deb9ecb2aeb0f775383bbd7b3d57

LCD Brightness - continued change to allow lcd dimming in both docked
and undocked modes.  It does not make sence to not allow screen to dim
in either instance.
Better options are welcome.

Change-Id: I501ec96fce64b9629a198ab713858eabee602600

allow LCD backlight decrease by default

Change-Id: I17616d17c86a2954d5c530c356b0541af5664ad5

auto-brightness - increase the wait to debounce light sensor values

  * needed for http://review.evervolv.com/#change,380
  * backlight changes too much without this fix
  * increase delay from 2 to 4 seconds

Change-Id: Icce6f34cbeccf479972603812d7df48e6dec6d95